Rename some files according to python convention rules
