+++ parameters/__init__.py
@@ -1,2 +1,2 @@
-from .Parser import Parser
-from .Rules import Rules
+from .rules import Rules
+from .parser import Parser
